checkNAN in sync
